Move `SessionProvider` to `shared/`
